requirements.txt
anyio==4.9.0
asgiref==3.9.1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
ecdsa==0.19.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jwcrypto==1.5.6
msgpack==1.1.1
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
PyJWT==2.10.1
pyparsing==3.2.3
python-jose==3.5.0
python-jwt==4.1.0
requests==2.32.4
requests-toolbelt==0.10.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==1.26.20
